combined server with backend class
